Added multiprocessing searching with URL address
