+++ scraper/test.py
@@ -26,30 +26,27 @@ import threading
 import time
 
 options = Options()
-#options.add_argument('--headless')
+options.add_argument('--headless')
 
 
 def search_neo24(search_for):
    driver = webdriver.Chrome(options=options)
-   driver.get("https://mediamarkt.pl")
    driver.set_window_size(1920, 1080)
-   input_element = driver.find_element_by_id("query_querystring")
+   driver.get("https://morele.net")
+   input_element = driver.find_element_by_xpath('//input[@name="search"]')
    input_element.send_keys(search_for)
    input_element.send_keys(Keys.ENTER)
    delay = 3  # seconds
    try:
       WebDriverWait(driver, delay).until(
-         EC.presence_of_all_elements_located((By.XPATH, '//*[@id="js-mainWrapper"]/main/div[6]/div[5]/div[2]/div')))
-      print("Page is ready! MM")
+         EC.presence_of_element_located((By.CLASS_NAME, 'cat-list-products')))
+      print("Page is ready!")
    except TimeoutException:
-      print("The parameter was not find! MM")
-      # print(driver.current_url)
-      url = driver.current_url
-      driver.close()
-      return url
+      print("The parameter was not find!")
+      return None
    page = driver.page_source
    driver.close()
-   # print("mm - done")
+   print("morele - done")
    return page
 
 
@@ -59,7 +56,7 @@ if __name__ == '__main__':
    #file = open("result.html", 'r')
    #data = file.read()
    #file.close()
-   parser.mediamarkt(search_neo24("Smartfon APPLE iPhone 11 Pro Max 64GB"), "Smartfon APPLE iPhone 11 Pro Max 64GB Złoty", data_dict)
+   parser.morele(search_neo24("Apple iPhone 11 Pro 64GB (srebrny)"), "Apple iPhone 11 Pro 64GB (srebrny)", data_dict)
    #scrap = UpcScrapper()
    #print(scrap.mediamarkt("https://mediamarkt.pl/telefony-i-smartfony/smartfon-apple-iphone-11-pro-max-64gb-zloty-mwhg2pm-a?querystring=Smartfon%20APPLE%20iPhone%2011%20Pro%20Max%2064GB%20Złoty"))
    print(data_dict)
