+++ scraper/webdrvr.py
@@ -11,83 +11,7 @@ from selenium.webdriver.common.by import By
 
 searchingPhrase = "50ep640"
 options = Options()
-#options.add_argument('--headless')
-
-def get_data_from_url(url):
-    scrap = UpcScrapper()
-    if 'euro.com.pl' in url:
-        return scrap.euro(url)
-    if 'morele.net' in url:
-        return scrap.morele(url)
-    if 'x-kom.pl' in url:
-        return scrap.xkom(url)
-    if 'mediaexpert.pl' in url:
-        return scrap.mediaexpert(url)
-    if 'mediamarkt.pl' in url:
-        return scrap.mediamarkt(url)
-    if 'komputronik.pl' in url:
-        return scrap.komputronik(url)
-    if 'neo24.pl' in url:
-        return scrap.neo24(url)
-
-def get_pages(name):
-    manager = multiprocessing.Manager()
-    data_dict = manager.dict()
-    euro = multiprocessing.Process(target=search_euro, args=(name, data_dict))
-    euro.start()
-    mediaexpert = multiprocessing.Process(target=search_mediaexpert, args=(name, data_dict))
-    mediaexpert.start()
-    mediamarkt = multiprocessing.Process(target=search_mediamarkt, args=(name, data_dict))
-    mediamarkt.start()
-    neo24 = multiprocessing.Process(target=search_neo24, args=(name, data_dict))
-    neo24.start()
-    morele = multiprocessing.Process(target=search_morele, args=(name, data_dict))
-    morele.start()
-    komputronik = multiprocessing.Process(target=search_komputronik, args=(name, data_dict))
-    komputronik.start()
-
-    euro.join()
-    mediaexpert.join()
-    mediamarkt.join()
-    neo24.join()
-    morele.join()
-    komputronik.join()
-    print("all data covered")
-    pages = data_dict
-    return pages
-
-
-def scrap_all_url(url):
-    data = get_data_from_url(url)
-    print(data['item'])
-    pages = get_pages(data['item'])
-    manager = multiprocessing.Manager()
-    data_dict = manager.dict()
-    print(str(pages['neo24']))
-    print("starting multithreaded parsing")
-    #print(pages)
-    euro = multiprocessing.Process(target=sp.euro, args=(str(pages['euro']), data['item'], data_dict))
-    euro.start()
-    mediaexpert = multiprocessing.Process(target=sp.mediaexpert, args=(str(pages['mediaexpert']), data['item'], data_dict))
-    mediaexpert.start()
-    mediamarkt = multiprocessing.Process(target=sp.mediamarkt, args=(str(pages['mediamarkt']), data['item'], data_dict))
-    mediamarkt.start()
-    neo24 = multiprocessing.Process(target=sp.neo24, args=(str(pages['neo24']), data['item'], data_dict))
-    neo24.start()
-    morele = multiprocessing.Process(target=sp.morele, args=(str(pages['morele']), data['item'], data_dict))
-    morele.start()
-    komputronik = multiprocessing.Process(target=sp.morele, args=(str(pages['komputronik']), data['item'], data_dict))
-    komputronik.start()
-
-    euro.join()
-    mediaexpert.join()
-    mediamarkt.join()
-    neo24.join()
-    morele.join()
-    komputronik.join()
-
-    return data_dict
-
+options.add_argument('--headless')
 
 def search_euro(search_for, return_dict):
     driver = webdriver.Chrome(options=options)
@@ -98,7 +22,7 @@ def search_euro(search_for, return_dict):
     input_element.send_keys(Keys.ENTER)
     page = driver.page_source
     driver.close()
-    print("euro - done")
+    #print("euro - done")
     return_dict['euro'] = page
 
 
@@ -109,14 +33,15 @@ def search_mediaexpert(search_for, return_dict):
    input_element = driver.find_element_by_css_selector('div.c-search_input').find_element_by_tag_name('input')
    input_element.send_keys(search_for)
    input_element.send_keys(Keys.ENTER)
-   delay = 3 # seconds
+   delay = 5 # seconds
    try:
       WebDriverWait(driver, delay).until(EC.presence_of_all_elements_located((By.XPATH, '/html/body/div[1]/div[13]/div[2]/div[5]/div[1]')))
-      print("Page is ready! ME")
+      #print("Page is ready! ME")
    except TimeoutException:
-      print("The parameter was not find! ME")
+      #print("The parameter was not find! ME")
       #print(driver.current_url)
       return_dict['mediaexpert'] = driver.current_url
+      driver.close()
       return
    page = driver.page_source
    driver.close()
@@ -130,13 +55,13 @@ def search_mediamarkt(search_for, return_dict):
     input_element = driver.find_element_by_id("query_querystring")
     input_element.send_keys(search_for)
     input_element.send_keys(Keys.ENTER)
-    delay = 3  # seconds
+    delay = 5  # seconds
     try:
         WebDriverWait(driver, delay).until(
             EC.presence_of_all_elements_located((By.XPATH, '//*[@id="js-mainWrapper"]/main/div[6]/div[5]/div[2]/div')))
-        print("Page is ready! MM")
+        #print("Page is ready! MM")
     except TimeoutException:
-        print("The parameter was not find! MM")
+        #print("The parameter was not find! MM")
         # print(driver.current_url)
         return_dict['mediamarkt'] = driver.current_url
         driver.close()
@@ -156,7 +81,7 @@ def search_xkom(search_for, return_dict):
     input_element.send_keys(Keys.ENTER)
     page = driver.page_source
     driver.close()
-    print("xkom - done")
+    #print("xkom - done")
     return page
 
 
@@ -167,12 +92,12 @@ def search_komputronik(search_for, return_dict):
     input_element = driver.find_element_by_xpath('//input[@type="text"]')
     input_element.send_keys(search_for)
     input_element.send_keys(Keys.ENTER)
-    delay = 3  # seconds
+    delay = 5  # seconds
     try:
         WebDriverWait(driver, delay).until(EC.presence_of_element_located((By.CLASS_NAME, 'product-entry2 ')))
-        print("Page is ready! Kom")
+        #print("Page is ready! Kom")
     except TimeoutException:
-        print("The parameter was not find!  Kom")
+        #print("The parameter was not find!  Kom")
         return_dict['komputronik'] = driver.current_url
         driver.close()
         return
@@ -188,34 +113,37 @@ def search_neo24(search_for, return_dict):
     input_element = driver.find_element_by_xpath('//input[@placeholder="Wpisz czego szukasz"]')
     input_element.send_keys(search_for)
     input_element.send_keys(Keys.ENTER)
-    delay = 10  # seconds
+    delay = 5  # seconds
     try:
         WebDriverWait(driver, delay).until(EC.presence_of_all_elements_located((By.ID, 'listingContent')))
-        print("Page is ready!")
+        #print("Page is ready!")
     except TimeoutException:
-        print("The parameter was not find! ")
+        #print("The parameter was not find! ")
         return_dict['neo24'] = driver.current_url
         driver.close()
         return
     page = driver.page_source
     driver.close()
-    print("neo - done")
+    #print("neo - done")
     return_dict['neo24'] = page
 
+
 def get_page_neo24(url):
-    driver = webdriver.Chrome(chrome_options=options)
+    driver = webdriver.Chrome(options=options)
     driver.set_window_size(1920, 1080)
     driver.get(url)
-    delay = 5  # seconds
+    delay = 3  # seconds
     try:
-        WebDriverWait(driver, delay).until(EC.presence_of_element_located((By.CLASS_NAME, 'productShopCss-neo24-product__price-12m')))
-        print("Page is ready!")
+        WebDriverWait(driver, delay).until(EC.presence_of_all_elements_located((By.CLASS_NAME, 'productShopCss-neo24-product__price-12m')))
+        #print("Page is ready!")
     except TimeoutException:
-        print("The parameter was not find!")
+        driver.close()
+        #print("The parameter was not find!")
         return None
     page = driver.page_source
     driver.close()
     return page
+
 
 def search_morele(search_for, return_dict):
     driver = webdriver.Chrome(options=options)
@@ -224,23 +152,17 @@ def search_morele(search_for, return_dict):
     input_element = driver.find_element_by_xpath('//input[@name="search"]')
     input_element.send_keys(search_for)
     input_element.send_keys(Keys.ENTER)
+    delay = 3  # seconds
+    try:
+        WebDriverWait(driver, delay).until(
+            EC.presence_of_element_located((By.CLASS_NAME, 'cat-list-products')))
+        print("Page is ready! Morle")
+    except TimeoutException:
+        print("The parameter was not find!")
+        driver.close()
+        return_dict['morele'] = None
     page = driver.page_source
     driver.close()
     print("morele - done")
     return_dict['morele'] = page
 
-
-
-if __name__ == "__main__":
-    #scrap = UpcScrapper()
-    #parser = UpcParser()
-    print(scrap_all_url("https://www.euro.com.pl/telefony-komorkowe/apple-iphone-pro-11-64gb-srebrny.bhtml"))
-    #print(search_mediamarkt("Smartfon APPLE iPhone 11 Pro Max 64GB Złoty"))
-    #print(scrap.morele("https://www.morele.net/sluchawki-steelseries-arctis-1-61427-5938473/"))
-    #print(scrap.komputronik("https://www.komputronik.pl/product/688817/huawei-matebook-x-pro-2020-green.html"))
-    #print(scrap.neo24Scraper(get_page_neo24('https://www.neo24.pl/philips-55-55pus6704-uhd.html')))
-    #print(scrap.euro("https://www.euro.com.pl/telewizory-led-lcd-plazmowe/tcl-50ep640-tv-led-4k-android.bhtml"))
-    #print(scrap.mediaexpert("https://www.mediaexpert.pl/telewizory-i-rtv/telewizory/telewizor-tcl-led-50ep680x1"))
-    #print(scrap.xkom("https://www.x-kom.pl/p/423390-narzedzie-serwisowe-sieciowe-phanteks-toolkit-zestaw-narzedzi.html"))
-    #print(scrap.neo24(get_page_neo24("https://www.neo24.pl/delonghi-odkamieniacz-ecodecalk-500ml.html"), "https://www.neo24.pl/delonghi-odkamieniacz-ecodecalk-500ml.html"))
-    #print(search_mediamarkt("aoc"))
